services/pi_frame_api/requirements.txt
fastapi
fastapi-utilities
python-multipart

uvicorn
apscheduler
pyyaml

RPi.GPIO
gpiozero
numpy
pillow
spidev

waveshare-epaper

pytest
httpx
